refactor: move removed-trademarks to unit test
Check sprite/costume library JSON data directly instead of searching
through the browser UI with Selenium. Runs in 0.5s vs 32.8s.

Refs #240
